+++ core/api_container/main.go
@@ -159,6 +159,7 @@ func createServiceNetworkAndModuleStore(
 		enclaveId,
 		dockerManager,
 		enclaveObjNameProvider,
+		enclaveObjLabelsProvider,
 		dockerNetworkId,
 		freeIpAddrTracker,
 		filesArtifactCache,