+++ core/commons/enclave_object_labels/enclave_object_labels.go
@@ -15,6 +15,10 @@ const (
 	// This is the static value that every single Kurtosis object will receive for the app ID label
 	AppIDValue = "kurtosis"
 
+	// TODO RENAME THIS!!! It's being used for more than just containers (also networks & volumes)
+	//  What we should really do though is have a separate type per object type (e.g. ContainerLabels, VolumeLabels, etc.)
+	//  and then just duplicate things like "enclave-id" that are duplicated, because even though they have the same value
+	//  they're actually entirely distinct, different labels
 	EnclaveIDContainerLabel = labelNamespace + "enclave-id"
 	ContainerTypeLabel      = labelNamespace + "container-type"
 
@@ -23,8 +27,10 @@ const (
 
 	// A label for the API container IP address that it's running on
 	APIContainerIPLabel    = labelNamespace + "api-container-ip"
-	// A label for the API container port that it's running on
-	APIContainerPortLabel    = labelNamespace + "api-container-port"
+	// Port number that the API container is listening on, INSIDE the network
+	APIContainerPortNumLabel = labelNamespace + "api-container-port-number"
+	// Protocol of the port that the API container is listening on
+	APIContainerPortProtocolLabel = labelNamespace + "api-container-port-protocol"
 
 	// Values for ContainerTypeLabel
 	ContainerTypeAPIContainer               = "api-container"
@@ -34,4 +40,10 @@ const (
 	ContainerTypeModuleContainer            = "module"
 	// This is a little weird to have here because  this is only used by the CLI (which depends on this repo)
 	ContainerTypeInteractiveREPL            = "interactive-repl"
+	ContainerTypeFilesArtifactExpander      = "files-artifact-expander"
+
+	// Testsuite type label + values
+	TestsuiteTypeLabelKey = labelNamespace + "testsuite-type"
+	TestsuiteTypeLabelValue_MetadataAcquisition = "metadata-acquisition"
+	TestsuiteTypeLabelValue_TestRunning = "test-running"
 )